fix some problem with random and the buffer
buffer was taking only positive input so my plot was not correct
random was created in a loop for -> was not a good idea

+++ neuron.hpp
@@ -80,7 +80,7 @@ class Neuron {
 	 * @param a poisson distribution
 	 * @return a [bool] to know if spike or not
 	 */
-	bool update(double I, int poisson);
+	bool update(double I, double poisson);
 	/**
 	 * @brief same update but without the poisson distribution for the googletest
 	 * @param an external current
